move downloader to openfe from openfecli
